fix(tokenizer): enter ident mode when appropriate
There was a bug brought up in https://github.com/alexfertel/bulloak/issues/54
where the tokenizer incorrectly enters identifier mode after seeing a `when/given`.

This means that `bulloak` failed to scaffold the following:

```tree
Foo
└── It reverts when X.
```

Because it saw a `.` after `when`, even though this is a perfectly valid
tree.

We fix this by checking that `when/given` only make the tokenizer enter
identifier mode if the last processed token was a branching token
(either `Corner` or `Tee` currently).